Lv1 코드 리팩토리
( 10 / 48 ) 완료

+++ Level1/N numbers spaced by X.cpp
@@ -8,9 +8,9 @@ using namespace std;
 vector<long long> solution(int x, int n) {
     vector<long long> answer(n);
 
-    for (int i = 0; i < n; ++i)
+    for (long long i = 0; i < n; ++i)
     {
-        answer[i] = static_cast <long long>(i + 1) * static_cast<long long>(x);
+        answer[i] = (i + 1) * x;
     }
 
     return answer;
